refactor: keep calculators and entrypoint functions together in state_tax.py
